plot functions added

+++ models/clustering.py
@@ -324,7 +324,7 @@ class ClusteringPipeline:
                                                  bidim_sentence_embeddings[:, 1].reshape(-1, 1).astype('float'),
                                                  np.array(self.training_sentences).reshape(-1, 1),
                                                  predictions.reshape(-1, 1),
-                                                 probs.reshape(-1, 1)
+                                                 probs.reshape(-1, 1).astype('float')
                                                  ], axis=1)
         df = pd.DataFrame(x1_x2_sentence_cluster, columns=['x1', 'x2', 'sentence', 'cluster', 'probabilities'])
         print(f"{df.head(5)}")
@@ -336,7 +336,7 @@ class ClusteringPipeline:
              bidim_sentence_embeddings_wo_outlies[:, 1].reshape(-1, 1).astype('float'),
              np.array(self.training_sentences).reshape(-1, 1),
              predictions_wo_outliers.reshape(-1, 1),
-             probs[predictions != 0].reshape(-1, 1)
+             probs[predictions != 0].reshape(-1, 1).astype('float')
              ], axis=1)
         df = pd.DataFrame(x1_x2_sentence_cluster_wo_outliers, columns=['x1', 'x2', 'sentence', 'cluster', 'probabilities'])
         scatter_with_sentences(df)
